Syncrepl session handling
