+++ Lib/ldap/syncrepl.py
@@ -4,341 +4,21 @@ ldap.syncrepl - for implementing syncrepl consumer (see RFC 4533)
 See https://www.python-ldap.org/ for project details.
 """
 
-from uuid import UUID
-
-# Imports from pyasn1
-from pyasn1.type import tag, namedtype, namedval, univ, constraint
-from pyasn1.codec.ber import encoder, decoder
-
 from ldap.pkginfo import __version__, __author__, __license__
-from ldap.controls import RequestControl, ResponseControl
 from ldap import RES_SEARCH_RESULT, RES_SEARCH_ENTRY, RES_INTERMEDIATE
+from ldap.controls.syncrepl import (
+    SyncRequestControl, SyncStateControl, SyncDoneControl,
+)
+from ldap.response import Response
+from ldap.extop.syncrepl import (
+    SyncInfoMessage,
+    SyncInfoNewCookie, SyncInfoRefreshPresent, SyncInfoRefreshDelete,
+    SyncInfoIDSet,
+)
 
 __all__ = [
     'SyncreplConsumer',
 ]
-
-
-class SyncUUID(univ.OctetString):
-    """
-    syncUUID ::= OCTET STRING (SIZE(16))
-    """
-    subtypeSpec = constraint.ValueSizeConstraint(16, 16)
-
-
-class SyncCookie(univ.OctetString):
-    """
-    syncCookie ::= OCTET STRING
-    """
-
-
-class SyncRequestMode(univ.Enumerated):
-    """
-           mode ENUMERATED {
-               -- 0 unused
-               refreshOnly       (1),
-               -- 2 reserved
-               refreshAndPersist (3)
-           },
-    """
-    namedValues = namedval.NamedValues(
-        ('refreshOnly', 1),
-        ('refreshAndPersist', 3)
-    )
-    subtypeSpec = univ.Enumerated.subtypeSpec + constraint.SingleValueConstraint(1, 3)
-
-
-class SyncRequestValue(univ.Sequence):
-    """
-       syncRequestValue ::= SEQUENCE {
-           mode ENUMERATED {
-               -- 0 unused
-               refreshOnly       (1),
-               -- 2 reserved
-               refreshAndPersist (3)
-           },
-           cookie     syncCookie OPTIONAL,
-           reloadHint BOOLEAN DEFAULT FALSE
-       }
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.NamedType('mode', SyncRequestMode()),
-        namedtype.OptionalNamedType('cookie', SyncCookie()),
-        namedtype.DefaultedNamedType('reloadHint', univ.Boolean(False))
-    )
-
-
-class SyncRequestControl(RequestControl):
-    """
-    The Sync Request Control is an LDAP Control [RFC4511] where the
-    controlType is the object identifier 1.3.6.1.4.1.4203.1.9.1.1 and the
-    controlValue, an OCTET STRING, contains a BER-encoded
-    syncRequestValue.  The criticality field is either TRUE or FALSE.
-    [..]
-    The Sync Request Control is only applicable to the SearchRequest
-    Message.
-    """
-    controlType = '1.3.6.1.4.1.4203.1.9.1.1'
-
-    def __init__(self, criticality=1, cookie=None, mode='refreshOnly', reloadHint=False):
-        self.criticality = criticality
-        self.cookie = cookie
-        self.mode = mode
-        self.reloadHint = reloadHint
-
-    def encodeControlValue(self):
-        rcv = SyncRequestValue()
-        rcv.setComponentByName('mode', SyncRequestMode(self.mode))
-        if self.cookie is not None:
-            rcv.setComponentByName('cookie', SyncCookie(self.cookie))
-        if self.reloadHint:
-            rcv.setComponentByName('reloadHint', univ.Boolean(self.reloadHint))
-        return encoder.encode(rcv)
-
-
-class SyncStateOp(univ.Enumerated):
-    """
-           state ENUMERATED {
-               present (0),
-               add (1),
-               modify (2),
-               delete (3)
-           },
-    """
-    namedValues = namedval.NamedValues(
-        ('present', 0),
-        ('add', 1),
-        ('modify', 2),
-        ('delete', 3)
-    )
-    subtypeSpec = univ.Enumerated.subtypeSpec + constraint.SingleValueConstraint(0, 1, 2, 3)
-
-
-class SyncStateValue(univ.Sequence):
-    """
-       syncStateValue ::= SEQUENCE {
-           state ENUMERATED {
-               present (0),
-               add (1),
-               modify (2),
-               delete (3)
-           },
-           entryUUID syncUUID,
-           cookie    syncCookie OPTIONAL
-       }
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.NamedType('state', SyncStateOp()),
-        namedtype.NamedType('entryUUID', SyncUUID()),
-        namedtype.OptionalNamedType('cookie', SyncCookie())
-    )
-
-
-class SyncStateControl(ResponseControl):
-    """
-    The Sync State Control is an LDAP Control [RFC4511] where the
-    controlType is the object identifier 1.3.6.1.4.1.4203.1.9.1.2 and the
-    controlValue, an OCTET STRING, contains a BER-encoded SyncStateValue.
-    The criticality is FALSE.
-    [..]
-    The Sync State Control is only applicable to SearchResultEntry and
-    SearchResultReference Messages.
-    """
-    controlType = '1.3.6.1.4.1.4203.1.9.1.2'
-    opnames = ('present', 'add', 'modify', 'delete')
-
-    def decodeControlValue(self, encodedControlValue):
-        d = decoder.decode(encodedControlValue, asn1Spec=SyncStateValue())
-        state = d[0].getComponentByName('state')
-        uuid = UUID(bytes=bytes(d[0].getComponentByName('entryUUID')))
-        cookie = d[0].getComponentByName('cookie')
-        if cookie is not None and cookie.hasValue():
-            self.cookie = str(cookie)
-        else:
-            self.cookie = None
-        self.state = self.__class__.opnames[int(state)]
-        self.entryUUID = str(uuid)
-
-
-class SyncDoneValue(univ.Sequence):
-    """
-       syncDoneValue ::= SEQUENCE {
-           cookie          syncCookie OPTIONAL,
-           refreshDeletes  BOOLEAN DEFAULT FALSE
-       }
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.OptionalNamedType('cookie', SyncCookie()),
-        namedtype.DefaultedNamedType('refreshDeletes', univ.Boolean(False))
-    )
-
-
-class SyncDoneControl(ResponseControl):
-    """
-    The Sync Done Control is an LDAP Control [RFC4511] where the
-    controlType is the object identifier 1.3.6.1.4.1.4203.1.9.1.3 and the
-    controlValue contains a BER-encoded syncDoneValue.  The criticality
-    is FALSE (and hence absent).
-    [..]
-    The Sync Done Control is only applicable to the SearchResultDone
-    Message.
-    """
-    controlType = '1.3.6.1.4.1.4203.1.9.1.3'
-
-    def decodeControlValue(self, encodedControlValue):
-        d = decoder.decode(encodedControlValue, asn1Spec=SyncDoneValue())
-        cookie = d[0].getComponentByName('cookie')
-        if cookie.hasValue():
-            self.cookie = str(cookie)
-        else:
-            self.cookie = None
-        refresh_deletes = d[0].getComponentByName('refreshDeletes')
-        if refresh_deletes.hasValue():
-            self.refreshDeletes = bool(refresh_deletes)
-        else:
-            self.refreshDeletes = None
-
-
-class RefreshDelete(univ.Sequence):
-    """
-           refreshDelete  [1] SEQUENCE {
-               cookie         syncCookie OPTIONAL,
-               refreshDone    BOOLEAN DEFAULT TRUE
-           },
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.OptionalNamedType('cookie', SyncCookie()),
-        namedtype.DefaultedNamedType('refreshDone', univ.Boolean(True))
-    )
-
-
-class RefreshPresent(univ.Sequence):
-    """
-           refreshPresent [2] SEQUENCE {
-               cookie         syncCookie OPTIONAL,
-               refreshDone    BOOLEAN DEFAULT TRUE
-           },
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.OptionalNamedType('cookie', SyncCookie()),
-        namedtype.DefaultedNamedType('refreshDone', univ.Boolean(True))
-    )
-
-
-class SyncUUIDs(univ.SetOf):
-    """
-    syncUUIDs      SET OF syncUUID
-    """
-    componentType = SyncUUID()
-
-
-class SyncIdSet(univ.Sequence):
-    """
-     syncIdSet      [3] SEQUENCE {
-         cookie         syncCookie OPTIONAL,
-         refreshDeletes BOOLEAN DEFAULT FALSE,
-         syncUUIDs      SET OF syncUUID
-     }
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.OptionalNamedType('cookie', SyncCookie()),
-        namedtype.DefaultedNamedType('refreshDeletes', univ.Boolean(False)),
-        namedtype.NamedType('syncUUIDs', SyncUUIDs())
-    )
-
-
-class SyncInfoValue(univ.Choice):
-    """
-       syncInfoValue ::= CHOICE {
-           newcookie      [0] syncCookie,
-           refreshDelete  [1] SEQUENCE {
-               cookie         syncCookie OPTIONAL,
-               refreshDone    BOOLEAN DEFAULT TRUE
-           },
-           refreshPresent [2] SEQUENCE {
-               cookie         syncCookie OPTIONAL,
-               refreshDone    BOOLEAN DEFAULT TRUE
-           },
-           syncIdSet      [3] SEQUENCE {
-               cookie         syncCookie OPTIONAL,
-               refreshDeletes BOOLEAN DEFAULT FALSE,
-               syncUUIDs      SET OF syncUUID
-           }
-       }
-    """
-    componentType = namedtype.NamedTypes(
-        namedtype.NamedType(
-            'newcookie',
-            SyncCookie().subtype(
-                implicitTag=tag.Tag(tag.tagClassContext, tag.tagFormatSimple, 0)
-            )
-        ),
-        namedtype.NamedType(
-            'refreshDelete',
-            RefreshDelete().subtype(
-                implicitTag=tag.Tag(tag.tagClassContext, tag.tagFormatSimple, 1)
-            )
-        ),
-        namedtype.NamedType(
-            'refreshPresent',
-            RefreshPresent().subtype(
-                implicitTag=tag.Tag(tag.tagClassContext, tag.tagFormatSimple, 2)
-            )
-        ),
-        namedtype.NamedType(
-            'syncIdSet',
-            SyncIdSet().subtype(
-                implicitTag=tag.Tag(tag.tagClassContext, tag.tagFormatSimple, 3)
-            )
-        )
-    )
-
-
-class SyncInfoMessage:
-    """
-    The Sync Info Message is an LDAP Intermediate Response Message
-    [RFC4511] where responseName is the object identifier
-    1.3.6.1.4.1.4203.1.9.1.4 and responseValue contains a BER-encoded
-    syncInfoValue.  The criticality is FALSE (and hence absent).
-    """
-    responseName = '1.3.6.1.4.1.4203.1.9.1.4'
-
-    def __init__(self, encodedMessage):
-        d = decoder.decode(encodedMessage, asn1Spec=SyncInfoValue())
-        self.newcookie = None
-        self.refreshDelete = None
-        self.refreshPresent = None
-        self.syncIdSet = None
-
-        # Due to the way pyasn1 works, refreshDelete and refreshPresent are both
-        # valid in the components as they are fully populated defaults. We must
-        # get the component directly from the message, not by iteration.
-        attr = d[0].getName()
-        comp = d[0].getComponent()
-
-        if comp is not None and comp.hasValue():
-            if attr == 'newcookie':
-                self.newcookie = str(comp)
-                return
-
-            val = {}
-
-            cookie = comp.getComponentByName('cookie')
-            if cookie.hasValue():
-                val['cookie'] = str(cookie)
-
-            if attr.startswith('refresh'):
-                val['refreshDone'] = bool(comp.getComponentByName('refreshDone'))
-            elif attr == 'syncIdSet':
-                uuids = []
-                ids = comp.getComponentByName('syncUUIDs')
-                for i in range(len(ids)):
-                    uuid = UUID(bytes=bytes(ids.getComponentByPosition(i)))
-                    uuids.append(str(uuid))
-                val['syncUUIDs'] = uuids
-                val['refreshDeletes'] = bool(comp.getComponentByName('refreshDeletes'))
-
-            setattr(self, attr, val)
 
 
 class SyncreplConsumer:
@@ -422,8 +102,9 @@ class SyncreplConsumer:
                 for m in msg:
                     dn, attrs, ctrls = m
                     for c in ctrls:
-                        if c.__class__.__name__ != 'SyncStateControl':
+                        if not isinstance(c, SyncStateControl):
                             continue
+
                         if c.state == 'present':
                             self.syncrepl_present([c.entryUUID])
                         elif c.state == 'delete':
@@ -432,40 +113,34 @@ class SyncreplConsumer:
                             self.syncrepl_entry(dn, attrs, c.entryUUID)
                             if self.__refreshDone is False:
                                 self.syncrepl_present([c.entryUUID])
+
                         if c.cookie is not None:
                             self.syncrepl_set_cookie(c.cookie)
                         break
 
             elif type == RES_INTERMEDIATE:
-                # Intermediate message. If it is a SyncInfoMessage, parse it
+                # Intermediate message, process any that are SyncInfoMessage
                 for m in msg:
-                    rname, resp, ctrls = m
-                    if rname != SyncInfoMessage.responseName:
-                        continue
-                    sim = SyncInfoMessage(resp)
-                    if sim.newcookie is not None:
-                        self.syncrepl_set_cookie(sim.newcookie)
-                    elif sim.refreshPresent is not None:
-                        self.syncrepl_present(None, refreshDeletes=False)
-                        if 'cookie' in sim.refreshPresent:
-                            self.syncrepl_set_cookie(sim.refreshPresent['cookie'])
-                        if sim.refreshPresent['refreshDone']:
+                    name, value, controls = m
+                    m = Response(mid, type, name=name, value=value,
+                                 controls=controls)
+                    if isinstance(m, SyncInfoNewCookie):
+                        self.syncrepl_set_cookie(m.cookie)
+                    elif isinstance(m, (SyncInfoRefreshPresent, SyncInfoRefreshDelete)):
+                        refreshDeletes = isinstance(m, SyncInfoRefreshDelete)
+                        self.syncrepl_present(None, refreshDeletes=refreshDeletes)
+                        if m.cookie is not None:
+                            self.syncrepl_set_cookie(m.cookie)
+                        if m.refreshDone:
                             self.__refreshDone = True
                             self.syncrepl_refreshdone()
-                    elif sim.refreshDelete is not None:
-                        self.syncrepl_present(None, refreshDeletes=True)
-                        if 'cookie' in sim.refreshDelete:
-                            self.syncrepl_set_cookie(sim.refreshDelete['cookie'])
-                        if sim.refreshDelete['refreshDone']:
-                            self.__refreshDone = True
-                            self.syncrepl_refreshdone()
-                    elif sim.syncIdSet is not None:
-                        if sim.syncIdSet['refreshDeletes'] is True:
-                            self.syncrepl_delete(sim.syncIdSet['syncUUIDs'])
+                    elif isinstance(m, SyncInfoIDSet):
+                        if m.refreshDeletes:
+                            self.syncrepl_delete(m.syncUUIDs)
                         else:
-                            self.syncrepl_present(sim.syncIdSet['syncUUIDs'])
-                        if 'cookie' in sim.syncIdSet:
-                            self.syncrepl_set_cookie(sim.syncIdSet['cookie'])
+                            self.syncrepl_present(m.syncUUIDs)
+                        if m.cookie is not None:
+                            self.syncrepl_set_cookie(m.cookie)
 
             if all == 0:
                 return True
